fixing validity period update fixing date format issue

+++ digiroad2-oracle/src/main/scala/fi/liikennevirasto/digiroad2/dao/MultiValueLinearAssetDao.scala
@@ -6,7 +6,7 @@ import fi.liikennevirasto.digiroad2.dao.Queries._
 import fi.liikennevirasto.digiroad2.linearasset.{MultiAssetValue, MultiValue, PersistedLinearAsset}
 import fi.liikennevirasto.digiroad2.oracle.MassQuery
 import org.joda.time.DateTime
-import org.joda.time.format.ISODateTimeFormat
+import org.joda.time.format.{DateTimeFormat, ISODateTimeFormat}
 import org.slf4j.LoggerFactory
 import slick.driver.JdbcDriver.backend.Database.dynamicSession
 import slick.jdbc.StaticQuery.interpolation
@@ -222,23 +222,24 @@ class MultiValueLinearAssetDao {
         }
 
       case Date =>
-        val formatter = ISODateTimeFormat.dateOptionalTimeParser()
-
+        val dateFormatter = DateTimeFormat.forPattern("dd.MM.yyyy")
         if (propertyValues.size > 1) throw new IllegalArgumentException("Date property must have exactly one value: " + propertyValues)
         if (propertyValues.isEmpty) {
           deleteDateProperty(assetId, propertyId).execute
         } else if (datePropertyValueDoesNotExist(assetId, propertyId)) {
-          insertDateProperty(assetId, propertyId, formatter.parseDateTime(propertyValues.head.value.toString)).execute
+          insertDateProperty(assetId, propertyId, dateFormatter.parseDateTime(propertyValues.head.value.toString)).execute
         } else {
-          updateDateProperty(assetId, propertyId, formatter.parseDateTime(propertyValues.head.value.toString)).execute
+          updateDateProperty(assetId, propertyId, dateFormatter.parseDateTime(propertyValues.head.value.toString)).execute
         }
 
       case TimePeriod =>
         //This type doesn't allow update
-        deleteValidityPeriodProperty(assetId, propertyId).execute
+        if(validityPeriodPropertyValueExist(assetId: Long, propertyId: Long)) {
+          deleteValidityPeriodProperty(assetId, propertyId).execute
+        }
 
         if (propertyValues.nonEmpty) {
-          propertyValues.map { propertyValue =>
+          propertyValues.distinct.foreach { propertyValue =>
             propertyValue.value match {
               case None =>
               case _ =>
@@ -261,6 +262,10 @@ class MultiValueLinearAssetDao {
 
   private def datePropertyValueDoesNotExist(assetId: Long, propertyId: Long) = {
     Q.query[(Long, Long), Long](existsDateProperty).apply((assetId, propertyId)).firstOption.isEmpty
+  }
+
+  private def validityPeriodPropertyValueExist(assetId: Long, propertyId: Long) = {
+    Q.query[(Long, Long), Long](existsValidityPeriodProperty).apply((assetId, propertyId)).firstOption.nonEmpty
   }
 
   private def textPropertyValueDoesNotExist(assetId: Long, propertyId: Long) = {
